requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cloudinary==1.42.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
Markdown==3.7
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
